+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.0.2
 
 - fixed active timestamp bug for closed planning nodes
+- use Myers diff algorithm for update functions
 
 ## 5.0.1